+++ crates/wdl-grammar/src/v1.rs
@@ -23,6 +23,7 @@
 //! | `missing_runtime_block` | `v1::W005` | Completeness | [Link](lint::MissingRuntimeBlock) |
 //! | `snake_case`            | `v1::W006` | Naming       | [Link](lint::SnakeCase)           |
 //! | `newline_eof`           | `v1::W007` | Spacing      | [Link](lint::NewlineEOF)          |
+//! | `preamble_comment`      | `v1::W010` | Style        | [Link](lint::PreambleComment)     |
 
 use pest::iterators::Pair;
 use pest::Parser as _;